edushare/requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==39.0.0
Django==3.2.16
django-colorfield==0.8.0
Pillow==9.5.0
pycparser==2.21
PyMySQL==1.0.2
pytz==2022.7
sqlparse==0.4.3
typing-extensions==4.4.0
